requirements.txt
fbchat
requests
requests-futures
tinydb
tabulate
nose
limp
markovify
textstat
pytz
beautifulsoup4
pyxdg
nltk
python-telegram-bot
